+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 * Fix #8118: bump knative.dev/networking from 0.0.0-20260529061306-c8de379770b9 to 0.0.0-20260921014731-e39778429a0f
 * Fix #8118: bump knative.dev/pkg from 0.0.0-20260422015212-ec452872dcc1 to 0.0.0-20260918182429-5dc1978f0042
 * Fix #8118: bump knative.dev/serving from 0.49.1 to 0.50.0
+* Fix #8126: bump prometheus-operator from 0.91.0 to 0.94.0
 * Fix #8100: bump tektoncd/pipeline from 1.12.0 to 1.16.0
 * Fix #8100: bump tektoncd/triggers from 0.36.0 to 0.37.0
 * Fix #8097: bump volcano.sh/apis from 1.11.0 to 1.15.2